fix: CI failures from 4.6.5 commit
Two issues from the previous commit broke CI:

1. Ruff RUF046: int(round(value)) is redundant — round() with no ndigits
   returns int already. Replaced with round(value).

2. Lingering TimerHandle: hass.loop.call_later in async_set_data_point
   leaked TimerHandles past test teardown, which pytest-homeassistant-
   custom-component flags as an error. Track the handle on the coordinator
   and cancel it in async_mark_destroyed and async_shutdown. Existing
   tests updated to call async_mark_destroyed in their finally blocks.

Bonus: rapid successive writes now cancel the previous pending refresh
instead of stacking timers.

Suite: 214/214 passing, ruff clean.

+++ custom_components/kkt_kolbe/sensor.py
@@ -181,7 +181,7 @@ class KKTKolbeSensor(KKTBaseEntity, SensorEntity):
         elif isinstance(value, float) and self._attr_suggested_display_precision == 0:
             # Tuya devices may report integer DPs as float (e.g. 10.0).
             # When precision=0, coerce to int so HA device view shows whole numbers.
-            self._cached_value = int(round(value))
+            self._cached_value = round(value)
         else:
             self._cached_value = value
